+++ NEWS.md
@@ -15,6 +15,7 @@
 * `and3s`, `or3s`, parallelized and _separated_ versions of `&`
 * `sum_and3s` and `sum_or3s`, the sums of the above logical vectors. 
 * `whichs` for an alterantive implementation of `which` which separates the input
+* `which_firstNA` and `which_lastNA` for first/last position of missing values
 
 ## Enhancements:
 * `which_first` accepts argument `use.which.max` for better performance on known short inputs